% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/explain.R
\name{explain.default}
\alias{explain.default}
\title{explain an S3 object in plain English}
\usage{
\method{explain}{default}(x, theme = "default", ...)
}
\arguments{
\item{x}{an object to be explained}

\item{theme}{theme to use}
}
\description{
explain an S3 object in plain English
}

