% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/explain.R
\name{explain}
\alias{explain}
\title{Explain an object in plain English}
\usage{
explain(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{extra arguments}
}
\value{
A text description of the object
}
\description{
Explain an object in plain English
}

