% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/explain_dplyr.R
\name{explain_line}
\alias{explain_line}
\title{explain a line of dplyr code}
\usage{
explain_line(expr, value, ok, visible)
}
\arguments{
\item{expr}{expression performed}

\item{value}{value returned}

\item{ok}{ok}

\item{visible}{visible}
}
\description{
explain a line of dplyr code
}

