% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/explainr_output.R
\name{explainr_output}
\alias{explainr_output}
\title{Wrap an object as the output of an \code{explain} function}
\usage{
explainr_output(x)
}
\arguments{
\item{x}{}
}
\description{
Wrap an object as the output of an \code{explain} function
}

