% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zzz.R
\name{explainr_themes}
\alias{explainr_themes}
\title{Built-in themes}
\description{
These are loaded from the \code{themes/} directory in the package, using
\code{\link{load_themes}}. So far they include:
\itemize{
  \item{default}{default text description}
  \item{visual}{graphical displays of objects}
}
}

