% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load.R
\name{load_template}
\alias{load_template}
\title{Load a .Rmd template for displaying an object}
\usage{
load_template(path)
}
\arguments{
\item{path}{Path to the .Rmd template file}
}
\value{
a list with two components:
  \item{header}{A named list containing the YAML components}
  \item{content}{A string with the contents of the string}
}
\description{
Load a .Rmd template for displaying an object
}

