% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load.R
\name{load_theme}
\alias{load_theme}
\title{Load a theme, which is a directory of templates}
\usage{
load_theme(path)
}
\arguments{
\item{path}{Path to a theme directory of .Rmd files}
}
\value{
a named list of templates
}
\description{
Load a theme, which is a directory of templates
}

