% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load.R
\name{load_themes}
\alias{load_themes}
\title{Load all themes}
\usage{
load_themes(directory = system.file("themes", package = "explainr"))
}
\arguments{
\item{directory}{directory to load themes from}
}
\value{
a named list of themes
}
\description{
Load themes, by default from explainr's built-in themes directory.
}

