% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/explainr_output.R
\name{print.explainr_output}
\alias{print.explainr_output}
\title{Print explainr output}
\usage{
\method{print}{explainr_output}(x)
}
\arguments{
\item{x}{}
}
\description{
Print explainr output
}

