% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/explain.R
\name{compile_template}
\alias{compile_template}
\title{compile an object using a template}
\usage{
compile_template(x, template)
}
\arguments{
\item{x}{object}

\item{template}{a template, typically loaded from \code{load_template}}
}
\value{
A compiled text version of the object
}
\description{
compile an object using a template
}

